app-catalog: Add comp and story for AppCatalogSettings
