+++ app-catalog/src/index.tsx
@@ -3,11 +3,11 @@ import {
   registerRoute,
   registerSidebarEntry,
 } from '@kinvolk/headlamp-plugin/lib';
+import { AppCatalogSettings } from '../src/components/settings/AppCatalogSettings';
 import ChartDetails from './components/charts/Details';
 import { ChartsList } from './components/charts/List';
 import ReleaseDetail from './components/releases/Detail';
 import ReleaseList from './components/releases/List';
-import { AppCatalogSettings } from './settings';
 
 export function isElectron(): boolean {
   // Renderer process
